refactor: move clients helper
