+++ go-chaos/backend/connection.go
@@ -132,20 +132,6 @@ func DisconnectBroker(disconnectBrokerCfg DisconnectBrokerCfg) error {
 	return disconnectPods(k8Client, broker1Pod, broker2Pod, disconnectBrokerCfg.OneDirection)
 }
 
-func ConnectToZeebeCluster(k8Client internal.K8Client) (zbc.Client, func(), error) {
-	port := 26500
-	closeFn := k8Client.MustGatewayPortForward(port, port)
-
-	zbClient, err := internal.CreateZeebeClient(port)
-	if err != nil {
-		return nil, closeFn, err
-	}
-	return zbClient, func() {
-		zbClient.Close()
-		closeFn()
-	}, nil
-}
-
 type DisconnectGatewayCfg struct {
 	DisconnectToAll bool
 	OneDirection    bool
